fix: 평균 점수 수정 - Float -> Double

+++ src/main/java/com/musicpedia/musicpediaapi/domain/rating/service/RatingService.java
@@ -49,7 +49,7 @@ public class RatingService {
         String ratingScore;
         String averageScore;
 
-        Float calculatedAverageScore = ratingRepository.calculateAverageScoreBySpotifyId(spotifyId);
+        Double calculatedAverageScore = ratingRepository.calculateAverageScoreBySpotifyId(spotifyId);
         if (calculatedAverageScore == null) {
             averageScore = "0";
         } else {
